Restructured Project folder
