+++ src/test/java/com/softwareonpurpose/indentmanager/IndentManagerTest.java
@@ -4,8 +4,6 @@ import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-import java.util.Locale;
-
 @Test
 public class IndentManagerTest {
     @DataProvider
